Refactored CurrentTheme import/export step.
